+++ src/app/providers/StoreProvider/index.ts
@@ -1,11 +1,14 @@
 import { StoreProvider } from './ui/StoreProvider';
 import { createReduxStore, AppDispatch } from './config/store';
-import type { StateSchema, ReduxStoreWithManager, ThunkConfig } from './config/StateSchema';
+import type {
+    StateSchema, ReduxStoreWithManager, ThunkConfig, StateSchemaKey,
+} from './config/StateSchema';
 
 export {
     StoreProvider,
     createReduxStore,
     StateSchema,
     ReduxStoreWithManager, ThunkConfig,
+    StateSchemaKey,
 };
 export type { AppDispatch };
