+++ src/app/providers/themeProvider/index.ts
@@ -1,5 +1,3 @@
 import ThemeProvider from './ui/ThemeProvider';
-import { useTheme } from './lib/useTheme';
-import { Theme } from './lib/ThemeContext';
 
-export { ThemeProvider, useTheme, Theme };
+export { ThemeProvider };
