+++ src/app/providers/themeProvider/ui/ThemeProvider.tsx
@@ -1,7 +1,7 @@
-import {
-    FC, ReactNode, useMemo, useState,
-} from 'react';
-import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from '../lib/ThemeContext';
+import { ReactNode, useMemo, useState } from 'react';
+import { LOCAL_STORAGE_THEME_KEY } from '../../../../shared/const/localstorage';
+import { Theme } from '../../../../shared/const/theme';
+import { ThemeContext } from '../../../../shared/lib/context/ThemeContext';
 
 const defaultTheme = localStorage.getItem(LOCAL_STORAGE_THEME_KEY) as Theme || Theme.LIGHT;
