+++ src/entities/Profile/ui/ProfileCard/ProfileCard.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import { ThemeDecorator } from '../../../../shared/config/storybook/ThemeDecorator';
-import { Theme } from '../../../../app/providers/themeProvider';
+import { Theme } from '../../../../shared/const/theme';
 import { ProfileCard } from './ProfileCard';
 import { Country } from '../../../../entities/Country';
 import { Currency } from '../../../../entities/Currency';
