+++ src/features/LangSwitcher/ui/LangSwitcher.module.scss
