+++ src/features/ThemeSwitcher/ui/ThemeSwitcher.tsx
@@ -1,9 +1,10 @@
 import { memo } from 'react';
-import { Theme, useTheme } from '../../../app/providers/themeProvider';
+import { useTheme } from '../../../shared/lib/hooks/useTheme/useTheme';
 import { classNames } from '../../../shared/lib/classNames/classNames';
 import LightIcon from '../../../shared/assets/icons/theme-light.svg';
 import DarkIcon from '../../../shared/assets/icons/theme-dark.svg';
 import { Button, ButtonTheme } from '../../../shared/ui/Button/Button';
+import { Theme } from '../../../shared/const/theme';
 
 interface ThemeSwitcherProps {
     className?: string;
