+++ src/pages/MainPage/ui/MainPage.stories.tsx
@@ -32,7 +32,7 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 import { ThemeDecorator } from '../../../shared/config/storybook/ThemeDecorator';
 import { StoreDecorator } from '../../../shared/config/storybook/StoreDecorator';
-import { Theme } from '../../../app/providers/themeProvider';
+import { Theme } from '../../../shared/const/theme';
 import MainPage from './MainPage';
 
 export default {
