+++ src/pages/NotFoundPage/ui/NotFoundPage.stories.tsx
@@ -30,7 +30,7 @@
 import React from 'react';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 import { ThemeDecorator } from '../../../shared/config/storybook/ThemeDecorator';
-import { Theme } from '../../../app/providers/themeProvider';
+import { Theme } from '../../../shared/const/theme';
 import { NotFoundPage } from './NotFoundPage';
 
 export default {
