+++ src/shared/config/storybook/ThemeDecorator.tsx
@@ -12,7 +12,8 @@
 /// ///////////////////////////////////////////////////////////////////////////////////////////
 
 import { Story } from '@storybook/react';
-import { Theme, ThemeProvider } from 'app/providers/themeProvider';
+import { ThemeProvider } from '../../../app/providers/themeProvider';
+import { Theme } from '../../../shared/const/theme';
 
 export const ThemeDecorator = (theme: Theme) => (StoryComponent: Story) => (
     <ThemeProvider initialTheme={theme}>
