+++ src/shared/const/localstorage.ts
@@ -1,2 +1,3 @@
 export const USER_LOCALSTORAGE_KEY = 'user';
 export const ARTICLES_VIEW_LOCALSTORAGE_KEY = 'articles_view';
+export const LOCAL_STORAGE_THEME_KEY = 'theme';
