+++ src/shared/lib/context/ThemeContext.ts
@@ -1,12 +1,5 @@
 import { createContext } from 'react';
-
-export const enum Theme {
-    LIGHT = 'app_light_theme',
-    DARK = 'app_dark_theme',
-    ORANGE = 'app_orange_theme',
-}
-
-export const LOCAL_STORAGE_THEME_KEY = 'theme';
+import { Theme } from '../../../shared/const/theme';
 
 export interface ThemeContextProps {
     theme?: Theme;
