+++ src/shared/lib/hooks/useTheme/useTheme.ts
@@ -1,6 +1,8 @@
 /* eslint-disable no-unused-expressions */
 import { useContext } from 'react';
-import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from '../lib/ThemeContext';
+import { LOCAL_STORAGE_THEME_KEY } from '../../../../shared/const/localstorage';
+import { Theme } from '../../../../shared/const/theme';
+import { ThemeContext } from '../../context/ThemeContext';
 
 interface UseThemeResult {
     toggleTheme: () => void,
