+++ src/shared/ui/Drawer/Drawer.tsx
@@ -2,11 +2,11 @@ import {
     memo, ReactNode, useCallback, useEffect,
 } from 'react';
 import { classNames } from '../../../shared/lib/classNames/classNames';
-import { useTheme } from '../../../app/providers/themeProvider';
 import { Overlay } from '../Overlay/Overlay';
 import cls from './Drawer.module.scss';
 import { Portal } from '../Portal/Portal';
 import { AnimationProvider, useAnimationLibs } from '../../../shared/lib/components/AnimationProvider';
+import { useTheme } from '../../../shared/lib/hooks/useTheme/useTheme';
 
 interface DrawerProps {
     className?: string;
