+++ src/shared/ui/Skeleton/Skeleton.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 import { ThemeDecorator } from '../../../shared/config/storybook/ThemeDecorator';
-import { Theme } from '../../../app/providers/themeProvider';
+import { Theme } from '../../../shared/const/theme';
 import { Skeleton } from './Skeleton';
 
 export default {
