+++ src/shared/ui/Text/Text.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 import { Text, TextSize, TextTheme } from '../../../shared/ui/Text/Text';
 import { ThemeDecorator } from '../../../shared/config/storybook/ThemeDecorator';
-import { Theme } from '../../../app/providers/themeProvider';
+import { Theme } from '../../../shared/const/theme';
 
 export default {
     title: 'shared/Text',
