+++ src/widgets/ErrorPage/ui/ErrorPage.stories.tsx
@@ -27,7 +27,7 @@
 // Dark.decorators = [ThemeDecorator(Theme.DARK)];
 
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { Theme } from '../../../app/providers/themeProvider';
+import { Theme } from '../../../shared/const/theme';
 import { ThemeDecorator } from '../../../shared/config/storybook/ThemeDecorator';
 import { ErrorPage } from './ErrorPage';
