+++ src/widgets/Navbar/ui/Navbar.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import { StoreDecorator } from '../../../shared/config/storybook/StoreDecorator';
 import { ThemeDecorator } from '../../../shared/config/storybook/ThemeDecorator';
-import { Theme } from '../../../app/providers/themeProvider';
+import { Theme } from '../../../shared/const/theme';
 import { Navbar } from './Navbar';
 
 export default {
