+++ src/widgets/Sidebar/ui/Sidebar/Sidebar.stories.tsx
@@ -1,5 +1,5 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { Theme } from '../../../../app/providers/themeProvider';
+import { Theme } from '../../../../shared/const/theme';
 import { StoreDecorator } from '../../../../shared/config/storybook/StoreDecorator';
 import { ThemeDecorator } from '../../../../shared/config/storybook/ThemeDecorator';
 import { Sidebar } from './Sidebar';
